requirements.txt
torchvision==0.2.1
numpy==1.15.0
torch==0.4.1
matplotlib==2.2.2
Pillow==5.2.0
